refactor: refactor jwt class
